refactor(print-format-builder): extract TableFieldInspector
